requirements.txt
aiohttp==3.8.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
blinker==1.6.2
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cleo==0.8.1
click==8.1.3
clikit==0.6.2
crashtest==0.3.1
cryptography==38.0.4
dataclasses-json==0.5.13
debugpy==1.6.5
distlib==0.3.6
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.104.1
filelock==3.9.0
Flask==2.2.2
Werkzeug==2.3.6
Flask-Cors==4.0.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
keyring==21.8.0
langchain==0.0.220
langchainplus-sdk==0.0.20
lockfile==0.12.2
MarkupSafe==2.1.2
marshmallow==3.20.1
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==1.0.0
ngrok==0.8.1
numexpr==2.8.4
numpy==1.24.1
openai==0.28.0
openapi-schema-pydantic==1.2.4
packaging==23.1
parso==0.8.3
passlib==1.7.4
pastel==0.2.1
pexpect==4.8.0
pg8000==1.30.2
pkginfo==1.9.6
platformdirs==3.8.0
pluggy==1.0.0
protobuf==4.21.12
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pycparser==2.21
pycryptodomex==3.16.0
pydantic==1.10.12
pydantic-core==2.10.1
pylev==1.4.0
pyodbc==4.0.39
pyparsing==3.0.9
pyseto==1.7.0
python-dotenv==1.0.0
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-multipart==0.0.6
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-toolbelt==0.9.1
scramp==1.4.4
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tenacity==8.2.2
tox==3.26.0
tqdm==4.60.0
typer==0.6.0
typing-extensions==4.8.0
ujson
urllib3==1.26.8
uvicorn==0.23.1
virtualenv==20.20.0
wrapt==1.13.3
xxhash==2.0.2
yarl==1.7.0
zipp==3.5.0
zope-event==4.5.0
zope-interface==5.4.0
azure-identity==1.7.1
azure-keyvault-secrets>=4.4.0,<4.6.0

